+++ src/main/java/team3176/robot/subsystems/superstructure/Superstructure.java
@@ -3,6 +3,9 @@ package team3176.robot.subsystems.superstructure;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import team3176.robot.subsystems.drivetrain.Drivetrain;
+import team3176.robot.subsystems.superstructure.elevator.Elevator;
+import team3176.robot.subsystems.superstructure.intake.Intake;
+import team3176.robot.subsystems.superstructure.shooter.Shooter;
 import team3176.robot.util.NoteVisualizer;
 
 public class Superstructure extends SubsystemBase {
@@ -20,81 +23,12 @@ public class Superstructure extends SubsystemBase {
     // transfer = Transfer.getInstance();
   }
 
-  public Command positiveIntake(double velocity) {
-    return this.run(() -> intake.setRollerVelocity(50))
-        .until(() -> intake.getIsRollerLinebreak())
-        .andThen(() -> intake.stopRoller());
-  }
-
-  public Command movePivotUp(double position) {
-    return this.run(() -> intake.setPivotPosition(.25)).andThen(() -> intake.stopPivot());
-  }
-
-  public Command movePivotDown(double position) {
-    return this.run(() -> intake.setPivotPosition(-position)).andThen(() -> intake.stopPivot());
-  }
-
-  public Command movePivotPostion(int position) {
-    return this.run(() -> intake.pivotGoToPosition(position));
-  }
-
-  public Command stopPivot() {
-    return this.run(() -> intake.stopPivot());
-  }
-
-  public Command negativeIntake(double velocity) {
-    return this.run(() -> intake.setRollerVelocity(-velocity));
-  }
-
-  public Command stopIntake() {
-    return this.run(() -> intake.stopRoller());
-  }
-
-  public Command moveElevator(double position) {
-    return this.run(() -> elevator.setElevatorMotor(.5));
-  }
-
-  public Command moveElevatorPID(int position) {
-    return this.run(() -> elevator.elevatorGoToPosition(position));
-  }
-
-  public Command stopElevator() {
-    return this.run(() -> elevator.stopElevator());
-  }
-
-  public Command upperShooterVelocity(double velocity) {
-    return this.run(() -> shooter.setUpperShooterVelocityVoltage(velocity));
-  }
-
-  public Command lowerShooterVelocity(double velocity) {
-    return this.run(() -> shooter.setLowerShooterVelocityVoltage(velocity));
-  }
-
-  public Command lowerShooterVelocity2(double velocity) {
-    return this.run(() -> shooter.setLowerShooterVelocityVoltage2(velocity));
-  }
-
   public Command runShooterPivot(double volts) {
     return shooter.pivotVoltage(volts);
   }
-  // public Command shooterPivot(double percent) {
-  //   return this.run(() -> shooter.setShooterPivotPercent(percent));
-  // }
-
-  // public Command stopShooterPivot() {
-  //   return this.run(() -> shooter.setShooterPivotPercent(0));
-  // }
-
-  // public Command stopShooterPivotPID() {
-  //   return this.run(() -> shooter.stopShooterPivotPosition());
-  // }
 
   public Command shooterPivotPID(int Position) {
     return shooter.pivotSetPositionOnce(Position);
-  }
-
-  public Command stopShooter() {
-    return this.run(() -> shooter.setShooterStop());
   }
 
   public static Superstructure getInstance() {
@@ -103,17 +37,4 @@ public class Superstructure extends SubsystemBase {
     }
     return instance;
   }
-
-  public Command kms() {
-    System.out.println("kms was calledf ##########################################");
-    return this.run(() -> shooter.debugger());
-  }
-
-  /*
-    public boolean getlinebreak1(){
-
-      return linebreak1.get();
-    }
-  /* */
-
 }
