+++ src/main/java/team3176/robot/subsystems/superstructure/intake/IntakeIOTalon.java
@@ -5,7 +5,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package team3176.robot.subsystems.superstructure;
+package team3176.robot.subsystems.superstructure.intake;
 
 import com.ctre.phoenix6.StatusCode;
 import com.ctre.phoenix6.configs.TalonFXConfiguration;
