+++ src/main/java/team3176/robot/subsystems/superstructure/transfer/Transfer.java
@@ -1,4 +1,4 @@
-package team3176.robot.subsystems.superstructure;
+package team3176.robot.subsystems.superstructure.transfer;
 
 import edu.wpi.first.math.MathUtil;
 import edu.wpi.first.math.controller.PIDController;
